+++ Sources/Storage/DelegatedStorage.swift
@@ -6,7 +6,7 @@ public typealias StorageErrorClosure = (Error) -> Void
 
 /// Class with the main `CryptoKit` logic.
 open class DelegatedStorage: Storage {
-    private let delegate: StorageDelegate?
+    public let delegate: StorageDelegate?
     private let symmetricKey: SymmetricKey?
     private let nonce: AES.GCM.Nonce?
     private let authenticationTag: Data?